Renames ServiceInstanceListSuppliers to ServiceInstanceListSupplierBuilder
Puts static builder() method on ServiceInstanceListSupplier.

Deprecates ServiceInstanceListSupplier.fixed(Environment) and replaces with fixed(serviceId).
