+++ docs/src/main/asciidoc/spring-cloud-commons.adoc
@@ -959,7 +959,7 @@ public class CustomLoadBalancerConfiguration {
 	@Bean
 	public ServiceInstanceListSupplier discoveryClientServiceInstanceListSupplier(
 			ConfigurableApplicationContext context) {
-		return ServiceInstanceListSuppliers.builder()
+		return ServiceInstanceListSupplier.builder()
 					.withDiscoveryClient()
 					.withHealthChecks()
 					.withCaching()
@@ -968,7 +968,7 @@ public class CustomLoadBalancerConfiguration {
 	}
 ----
 
-TIP:: In order to make working on your own LoadBalancer configuration easier, we have added some utility methods in `ServiceInstanceListSuppliers` class.
+TIP:: In order to make working on your own LoadBalancer configuration easier, we have added a `builder()` method to the `ServiceInstanceListSupplier` class.
 
 TIP:: You can also use our alternative predefined configurations in place of the default ones by setting the value of `spring.cloud.loadbalancer.configurations` property to `zone-preference` to use `ZonePreferenceServiceInstanceListSupplier` with caching or to `health-check` to use `HealthCheckServiceInstanceListSupplier` with caching.
