+++ spring-cloud-loadbalancer/src/main/java/org/springframework/cloud/loadbalancer/core/ServiceInstanceListSupplier.java
@@ -39,8 +39,16 @@ public interface ServiceInstanceListSupplier
 
 	String getServiceId();
 
+	static ServiceInstanceListSupplierBuilder builder() {
+		return new ServiceInstanceListSupplierBuilder();
+	}
+
 	static FixedServiceInstanceListSupplier.Builder fixed(Environment environment) {
 		return new FixedServiceInstanceListSupplier.Builder(environment);
+	}
+
+	static FixedServiceInstanceListSupplier.SimpleBuilder fixed(String serviceId) {
+		return new FixedServiceInstanceListSupplier.SimpleBuilder(serviceId);
 	}
 
 	class FixedServiceInstanceListSupplier implements ServiceInstanceListSupplier {
@@ -49,13 +57,14 @@ public interface ServiceInstanceListSupplier
 
 		private List<ServiceInstance> instances;
 
+		@Deprecated
 		public static Builder with(Environment env) {
 			return new Builder(env);
 		}
 
-		private FixedServiceInstanceListSupplier(Environment env,
+		private FixedServiceInstanceListSupplier(String serviceId,
 				List<ServiceInstance> instances) {
-			this.serviceId = env.getProperty(PROPERTY_NAME);
+			this.serviceId = serviceId;
 			this.instances = instances;
 		}
 
@@ -69,6 +78,43 @@ public interface ServiceInstanceListSupplier
 			return Flux.just(instances);
 		}
 
+		@Deprecated
+		public static final class SimpleBuilder {
+
+			private final ArrayList<ServiceInstance> instances = new ArrayList<>();
+
+			private final String serviceId;
+
+			private SimpleBuilder(String serviceId) {
+				this.serviceId = serviceId;
+			}
+
+			public SimpleBuilder instance(ServiceInstance instance) {
+				instances.add(instance);
+				return this;
+			}
+
+			public SimpleBuilder instance(int port) {
+				return instance("localhost", port);
+			}
+
+			public SimpleBuilder instance(String host, int port) {
+				DefaultServiceInstance instance = new DefaultServiceInstance(
+						instanceId(serviceId, host, port), serviceId, host, port, false);
+				return instance(instance);
+			}
+
+			private String instanceId(String serviceId, String host, int port) {
+				return serviceId + ":" + host + ":" + port;
+			}
+
+			public FixedServiceInstanceListSupplier build() {
+				return new FixedServiceInstanceListSupplier(serviceId, instances);
+			}
+
+		}
+
+		@Deprecated
 		public static final class Builder {
 
 			private final Environment env;
@@ -99,7 +145,8 @@ public interface ServiceInstanceListSupplier
 			}
 
 			public FixedServiceInstanceListSupplier build() {
-				return new FixedServiceInstanceListSupplier(env, instances);
+				return new FixedServiceInstanceListSupplier(
+						env.getProperty(PROPERTY_NAME), instances);
 			}
 
 		}
